+++ backend/types.ts
@@ -94,199 +94,3 @@ export interface MeetingTime {
   end: number,
 }
 
-
-export type EsMapping = Record<string, any>;
-// TODO value is a Document--something we insert into ES
-export type EsBulkData = Record<string, any>;
-
-// ======== Results ==========
-// TODO blocked a bit by new ORM, trying to avoid duplication
-export type EsResult = any;
-
-export type EsMultiResult = {
-  body: {
-    responses: EsResultBody[]
-  }
-};
-
-export interface EsResultBody {
-  took: number,
-  hits: {
-    total: { value: number },
-    hits: EsResult[],
-    value: number,
-  }
-  aggregations: EsAggResults
-}
-
-export interface EsAggResults {
-  [aggName: string]: {
-    buckets: Array<{ key: string, doc_count: number }>
-  }
-}
-
-export interface SearchResults {
-  searchContent: SearchResult[],
-  resultCount: number,
-  took: {
-    total: number,
-    hydrate: number,
-    es: number,
-  },
-  aggregations: AggResults
-}
-
-export interface PartialResults {
-  output: EsResult[],
-  resultCount: number,
-  took: number,
-  aggregations: AggResults
-}
-
-export interface AggCount {
-  value: string;
-  count: number;
-}
-
-export interface AggResults {
-  [filterName: string]: AggCount[];
-}
-
-
-// WARNING there should be a `type` property of type `string` here, but Typescript doesn't like it (because `type`, the property name, is a string, which conflicts with the `resultId` field)
-export interface SearchResult {
-  [resultId: string]: ResultDoc
-}
-
-export interface CourseDoc {
-  class: Course;
-  sections: Section[];
-}
-export type ResultDoc = Employee | CourseDoc;
-
-// ======== Queries ==========
-export interface EsQuery {
-  from: number,
-  size: number,
-  sort: SortQuery,
-  query: QueryNode,
-  aggregations?: QueryAgg
-};
-
-// every query is either a leaf query or a bool query
-export type QueryNode = BoolQuery | LeafQuery;
-
-export interface BoolQuery {
-  bool: BoolType;
-};
-
-export type BoolType = MustQuery | ShouldQuery | FilterQuery;
-
-// THESE ARE THE SAME. The only difference is what string shows up first.
-export interface MustQuery {
-  must: OneOrMany<QueryNode>;
-};
-
-export interface ShouldQuery {
-  should: OneOrMany<QueryNode>;
-};
-
-export interface FilterQuery {
-  filter: OneOrMany<QueryNode>;
-};
-
-export type LeafQuery = TermQuery 
-                      | TermsQuery 
-                      | ExistsQuery 
-                      | MultiMatchQuery 
-                      | RangeQuery
-                      | typeof MATCH_ALL_QUERY;
-
-export interface TermQuery {
-  term: FieldQuery
-}
-
-export interface TermsQuery {
-  terms: FieldQuery
-}
-
-export interface MultiMatchQuery {
-  multi_match: {
-    query: string,
-    type: string,
-    fields: string[]
-  }
-}
-
-export interface RangeQuery {
-  range: {
-    [fieldName: string]: {
-      gte?: number;
-      gt?:  number;
-      lte?: number;
-      lt?:  number;
-    }
-  }
-}
-
-export const MATCH_ALL_QUERY = { match_all: {} };
-
-export interface ExistsQuery {
-  exists: { field: string };
-}
-
-// don't like the any
-export interface FieldQuery {
-  [fieldName: string]: any;
-}
-
-export type SortQuery = [string, SortStruct];
-
-// I hate this name
-export interface SortStruct {
-  [fieldName: string]: {
-    order: string,
-    unmapped_type: string
-  }
-}
-
-export interface QueryAgg {
-  [aggName: string]: {
-    terms: {
-      field: string,
-      size?: number,
-    }
-  }
-}
-
-type OneOrMany<T> = T | T[];
-
-// ========= Filters =========
-
-export type FilterInput = {
-  [filterName: string]: ValidFilterInput
-}
-
-export type AggProp = false | string; 
-
-export interface FilterStruct<Input> {
-  validate: (input: Input) => boolean,
-  create: (input: Input) => LeafQuery,
-  agg: AggProp;
-}
-
-export interface AggFilterStruct<Input> extends FilterStruct<Input> {
-  agg: string;
-}
-
-export interface Range {
-  min: number;
-  max: number;
-}
-
-export type ValidFilterInput = string | string[] | boolean | Range;
-
-export type FilterPrelude = Record<string, FilterStruct<ValidFilterInput>>;
-
-export type AggFilterPrelude = Record<string, AggFilterStruct<ValidFilterInput>>;
-
